backend/requirements.txt
# Core FastAPI and Web Framework
fastapi==0.109.2
uvicorn==0.27.1
starlette==0.36.3
pydantic==2.6.1
pydantic_core==2.16.2

# HTTP Client Libraries
httpx==0.28.1
aiohttp==3.12.0
aiohappyeyeballs==2.6.1
aiosignal==1.3.2
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
httpcore==1.0.9
httptools==0.6.4
httplib2==0.22.0
h11==0.16.0

# Database and ORM
SQLAlchemy==2.0.41
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis and Caching
redis==5.0.1

# Async Support
anyio==4.9.0
asgiref==3.8.1
sniffio==1.3.1

# Security and Authentication
cryptography==45.0.2
bcrypt==4.3.0
python-jose==3.3.0
ecdsa==0.19.1
oauthlib==3.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodome==3.23.0
cffi==1.17.1
pycparser==2.22

# Google Cloud Services
google-ai-generativelanguage==0.6.6
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.42.1
google-cloud-appengine-logging==1.6.1
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.33.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.14.0
google-crc32c==1.7.1
google-generativeai==0.7.2
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.62.3

# Vector Databases
chromadb==0.4.22
chroma-hnswlib==0.7.3
pinecone==7.0.1
pinecone-client==3.0.0
pinecone-plugin-interface==0.0.7

# Machine Learning and NLP
torch==2.2.0
transformers==4.37.2
tokenizers==0.15.2
safetensors==0.5.3
huggingface-hub==0.32.0
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
blis==1.2.1
catalogue==2.0.10
confection==0.1.5
cymem==2.0.11
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
murmurhash==1.0.13
preshed==3.0.9
srsly==2.5.1
thinc==8.3.4
typer==0.15.4
wasabi==1.1.3
weasel==0.4.1

# LangChain and RAG
langchain==0.2.17
langchain-core==0.2.43
langchain-google-genai==1.0.10
langchain-text-splitters==0.2.4
langsmith==0.1.147

# Scientific Computing
numpy==1.26.4
mpmath==1.3.0
sympy==1.14.0
networkx==3.4.2

# Privacy and PII Detection
presidio-analyzer==2.2.35
presidio-anonymizer==2.2.35
phonenumbers==8.13.55

# Voice and Scheduling (New for Blueprint Architecture)
croniter==1.4.1

# Crisis Intervention Communication
twilio==8.13.0                    # SMS and voice calls (fallback)
sendgrid==6.11.0                  # Email via SendGrid (fallback)
boto3==1.35.0                     # AWS SES email fallback

# Observability and Monitoring
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
posthog==4.2.0

# Storage and File Handling
cloudpathlib==0.21.1
fsspec==2025.5.1
smart-open==7.1.0
hf-xet==1.1.2
filelock==3.18.0

# Utilities
click==8.1.8
coloredlogs==15.0.1
humanfriendly==10.0
rich==14.0.0
tqdm==4.67.1
tenacity==8.5.0
backoff==2.2.1
overrides==7.7.0
shellingham==1.5.4
regex==2024.11.6
cachetools==5.5.2
durationpy==0.10
mmh3==5.1.0

# Data Validation and Serialization
attrs==25.3.0
orjson==3.10.18
PyYAML==6.0.2
jsonpatch==1.33
jsonpointer==3.0.0
python-multipart==0.0.9

# Date and Time
python-dateutil==2.9.0.post0

# Environment and Configuration
python-dotenv==1.0.0

# Text Processing and Markup
Jinja2==3.1.6
MarkupSafe==3.0.2
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.19.1

# Web Standards and Parsing
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
urllib3==2.4.0
uritemplate==4.1.1
tldextract==5.3.0

# Packaging and Build
build==1.2.2.post1
packaging==24.2
pyparsing==3.2.3
pyproject_hooks==1.2.0
setuptools==80.8.0
six==1.17.0
wheel

# Type System and Inspection
annotated-types==0.7.0
typing_extensions==4.13.2
typing-inspection==0.4.1

# Async and Concurrency
frozenlist==1.6.0
multidict==6.4.4
propcache==0.3.1
yarl==1.20.0
uvloop==0.21.0

# Geometry and Spatial
shapely==2.1.1

# Protocol Buffers and Serialization
protobuf==4.25.7
proto-plus==1.26.1
flatbuffers==25.2.10

# Statistics and Quality
Deprecated==1.2.18
distro==1.9.0

# WebSockets and Real-time
websockets==15.0.1
websocket-client==1.8.0
watchfiles==1.0.5

# Cloud and Kubernetes
kubernetes==32.0.1

# Message Queue and Streaming
pulsar-client==3.7.0

# ONNX Runtime for ML Inference
onnxruntime==1.22.0

# Database Query Builder
PyPika==0.48.9

# Security Token Libraries
rsa==4.9.1

# Metadata and Imports
importlib_metadata==8.4.0
importlib_resources==6.5.2
zipp==3.21.0

# Wrapper and Instrumentation
wrapt==1.17.2

# ML Enhancement Dependencies (from requirements-ml.txt)
# transformers>=4.21.0  # Already included above with specific version
# torch>=1.12.0  # Already included above with specific version

# Logging and Structured Logging
structlog==24.1.0

#image generation
pillow==11.1.0